backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
spacy==3.7.2
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
torch==2.1.0
scikit-learn==1.3.2